+++ cli/src/account.rs
@@ -10,7 +10,7 @@ pub fn data(client: &Client, matches: &ArgMatches) -> Result<()> {
     let endpoint = account::Data::new(id, key);
     let account = client.request(endpoint)?;
 
-    Formatter::start_stdout(Simple).render(&account);
+    Formatter::start_stdout(Simple::new()).render(&account);
 
     Ok(())
 }
@@ -20,7 +20,7 @@ pub fn details(client: &Client, matches: &ArgMatches) -> Result<()> {
     let endpoint = account::Details::new(id);
     let account = client.request(endpoint)?;
 
-    Formatter::start_stdout(Simple).render(&account);
+    Formatter::start_stdout(Simple::new()).render(&account);
 
     Ok(())
 }
@@ -37,7 +37,7 @@ pub fn transactions(client: &Client, matches: &ArgMatches) -> Result<()> {
     let iter = sync::Iter::new(&client, endpoint);
 
     let mut res = Ok(());
-    let mut fmt = Formatter::start_stdout(Simple);
+    let mut fmt = Formatter::start_stdout(Simple::new());
     pager.paginate(iter, |result| match result {
         Ok(txn) => fmt.render(&txn),
         Err(err) => res = Err(err.into()),
@@ -58,7 +58,7 @@ pub fn effects(client: &Client, matches: &ArgMatches) -> Result<()> {
     let iter = sync::Iter::new(&client, endpoint);
 
     let mut res = Ok(());
-    let mut fmt = Formatter::start_stdout(Simple);
+    let mut fmt = Formatter::start_stdout(Simple::new());
     pager.paginate(iter, |result| match result {
         Ok(effect) => fmt.render(&effect),
         Err(err) => res = Err(err.into()),
@@ -79,11 +79,31 @@ pub fn offers(client: &Client, matches: &ArgMatches) -> Result<()> {
     let iter = sync::Iter::new(&client, endpoint);
 
     let mut res = Ok(());
-    let mut fmt = Formatter::start_stdout(Simple);
+    let mut fmt = Formatter::start_stdout(Simple::new());
     pager.paginate(iter, |result| match result {
         Ok(offer) => fmt.render(&offer),
         Err(err) => res = Err(err.into()),
     });
     let _ = fmt.stop();
+    res
+}
+
+pub fn operations(client: &Client, matches: &ArgMatches) -> Result<()> {
+    let pager = Pager::from_arg(&matches);
+
+    let id = matches.value_of("ID").expect("ID is required");
+    let endpoint = account::Operations::new(id);
+    let endpoint = pager.assign(endpoint);
+    let endpoint = cursor::assign_from_arg(matches, endpoint);
+    let endpoint = ordering::assign_from_arg(matches, endpoint);
+
+    let iter = sync::Iter::new(&client, endpoint);
+
+    let mut res = Ok(());
+    let mut fmt = Formatter::start_stdout(Simple::new());
+    pager.paginate(iter, |result| match result {
+        Ok(op) => fmt.render(&op),
+        Err(err) => res = Err(err.into()),
+    });
     res
 }