+++ cli/src/fmt/simple/asset.rs
@@ -1,21 +1,42 @@
 use super::Simple;
 use fmt::Render;
-use stellar_client::resources::Asset;
+use stellar_client::resources::{Asset, AssetIdentifier, Flags};
 
 impl Render<Asset> for Simple {
     fn render(&self, asset: &Asset) -> Option<String> {
         let mut buf = String::new();
-        append_to_buffer!(buf, "Code:         {}", asset.code());
-        append_to_buffer!(buf, "Type:         {}", asset.asset_type());
-        append_to_buffer!(buf, "Issuer:       {}", asset.issuer());
-        append_to_buffer!(buf, "Amount:       {}", asset.amount());
-        append_to_buffer!(buf, "Num accounts: {}", asset.num_accounts());
-        append_to_buffer!(buf, "Flags:");
-        if asset.is_auth_required() {
-            append_to_buffer!(buf, "  auth is required");
+        append!(
+            buf,
+            "Asset:        {}",
+            self.render(asset.identifier()).unwrap()
+        );
+        append!(buf, "Amount:       {}", asset.amount());
+        append!(buf, "Num Accounts: {}", asset.num_accounts());
+        append!(buf, "Flags:");
+        nest!(buf, self, &asset.flags());
+        Some(buf)
+    }
+}
+
+impl Render<AssetIdentifier> for Simple {
+    fn render(&self, asset_id: &AssetIdentifier) -> Option<String> {
+        let asset_str = if asset_id.is_native() {
+            asset_id.code().to_string()
+        } else {
+            format!("{}-{}", asset_id.code(), asset_id.issuer())
+        };
+        Some(asset_str)
+    }
+}
+
+impl Render<Flags> for Simple {
+    fn render(&self, flags: &Flags) -> Option<String> {
+        let mut buf = String::new();
+        if flags.is_auth_required() {
+            append!(buf, "auth is required");
         }
-        if asset.is_auth_revocable() {
-            append_to_buffer!(buf, "  auth is revocable");
+        if flags.is_auth_revocable() {
+            append!(buf, "auth is revocable");
         }
         Some(buf)
     }