+++ cli/src/fmt/simple/datum.rs
@@ -5,7 +5,7 @@ use stellar_client::resources::Datum;
 impl Render<Datum> for Simple {
     fn render(&self, datum: &Datum) -> Option<String> {
         let mut buf = String::new();
-        append_to_buffer!(buf, "value:       {}", datum.value());
+        append!(buf, "value:       {}", datum.value());
         Some(buf)
     }
 }