+++ cli/src/fmt/simple/trade_aggregation.rs
@@ -5,15 +5,15 @@ use stellar_client::resources::TradeAggregation;
 impl Render<TradeAggregation> for Simple {
     fn render(&self, aggregation: &TradeAggregation) -> Option<String> {
         let mut buf = String::new();
-        append_to_buffer!(buf, "timestamp:      {}", aggregation.started_at());
-        append_to_buffer!(buf, "trade_count:    {}", aggregation.count());
-        append_to_buffer!(buf, "base_volume:    {}", aggregation.base_volume());
-        append_to_buffer!(buf, "counter_volume: {}", aggregation.counter_volume());
-        append_to_buffer!(buf, "average:        {}", aggregation.average());
-        append_to_buffer!(buf, "high:           {}", aggregation.high());
-        append_to_buffer!(buf, "low:            {}", aggregation.low());
-        append_to_buffer!(buf, "open:           {}", aggregation.open());
-        append_to_buffer!(buf, "close:          {}", aggregation.close());
+        append!(buf, "timestamp:      {}", aggregation.started_at());
+        append!(buf, "trade_count:    {}", aggregation.count());
+        append!(buf, "base_volume:    {}", aggregation.base_volume());
+        append!(buf, "counter_volume: {}", aggregation.counter_volume());
+        append!(buf, "average:        {}", aggregation.average());
+        append!(buf, "high:           {}", aggregation.high());
+        append!(buf, "low:            {}", aggregation.low());
+        append!(buf, "open:           {}", aggregation.open());
+        append!(buf, "close:          {}", aggregation.close());
         Some(buf)
     }
 }