+++ cli/src/payments.rs
@@ -15,7 +15,7 @@ pub fn all(client: &Client, matches: &ArgMatches) -> Result<()> {
     let iter = sync::Iter::new(&client, endpoint);
 
     let mut res = Ok(());
-    let mut fmt = Formatter::start_stdout(Simple);
+    let mut fmt = Formatter::start_stdout(Simple::new());
     pager.paginate(iter, |result| match result {
         Ok(operation) => fmt.render(&operation),
         Err(err) => res = Err(err.into()),