+++ cli/src/trades.rs
@@ -101,7 +101,7 @@ pub fn aggregations(client: &Client, matches: &ArgMatches) -> Result<()> {
     let iter = sync::Iter::new(&client, endpoint);
 
     let mut res = Ok(());
-    let mut fmt = Formatter::start_stdout(Simple);
+    let mut fmt = Formatter::start_stdout(Simple::new());
     pager.paginate(iter, |result| match result {
         Ok(aggregation) => fmt.render(&aggregation),
         Err(err) => res = Err(err.into()),